csv2word/requirements.txt
lxml==4.9.3
numpy==1.25.2
pandas==2.1.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
six==1.16.0
tzdata==2023.3
